chore: Code-split CS config class
